chore(Background): Cleanup
# Conflicts:
#	data/vehicle/src/main/sqldelight/4.db

+++ feature/main/src/main/kotlin/com/masselis/tpmsadvanced/feature/main/interfaces/composable/DeleteVehicleButtonTags.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.masselis.tpmsadvanced.core.ui.LocalHomeNavController
+import com.masselis.tpmsadvanced.core.ui.viewModel
 import com.masselis.tpmsadvanced.data.vehicle.model.Vehicle
 import com.masselis.tpmsadvanced.feature.main.R
 import com.masselis.tpmsadvanced.feature.main.interfaces.composable.DeleteVehicleButtonTags.Button.tag
@@ -31,13 +32,13 @@ import com.masselis.tpmsadvanced.feature.main.interfaces.viewmodel.DeleteVehicle
 import com.masselis.tpmsadvanced.feature.main.interfaces.viewmodel.DeleteVehicleViewModel.Event
 import com.masselis.tpmsadvanced.feature.main.interfaces.viewmodel.DeleteVehicleViewModel.State
 import com.masselis.tpmsadvanced.feature.main.ioc.vehicle.InternalVehicleComponent
-import com.masselis.tpmsadvanced.feature.main.ioc.vehicle.InternalVehicleComponent.Companion.viewModel
+import com.masselis.tpmsadvanced.feature.main.ioc.vehicle.VehicleComponent.Companion.key
 
 @Composable
 internal fun DeleteVehicleButton(
     modifier: Modifier = Modifier,
-    vehicleComponent: InternalVehicleComponent = LocalInternalVehicleComponent.current,
-    viewModel: DeleteVehicleViewModel = vehicleComponent.viewModel { it.DeleteVehicleViewModel() }
+    component: InternalVehicleComponent = LocalInternalVehicleComponent.current,
+    viewModel: DeleteVehicleViewModel = component.viewModel(component.key()) { it.DeleteVehicleViewModel() }
 ) {
     val navController = LocalHomeNavController.current
     val state by viewModel.stateFlow.collectAsState()
@@ -55,12 +56,10 @@ internal fun DeleteVehicleButton(
             Text(text = "Delete \"${state.vehicle.name}\"")
         }
     }
-    if (showDeleteDialog)
-        DeleteVehicleDialog(
-            vehicle = state.vehicle,
-            onDismissRequest = { showDeleteDialog = false },
-            onDelete = { viewModel.delete(); showDeleteDialog = false; }
-        )
+    if (showDeleteDialog) DeleteVehicleDialog(
+        vehicle = state.vehicle,
+        onDismissRequest = { showDeleteDialog = false },
+        onDelete = { viewModel.delete(); showDeleteDialog = false; })
     LaunchedEffect(viewModel) {
         for (event in viewModel.eventChannel) {
             when (event) {
@@ -80,23 +79,19 @@ private fun DeleteVehicleDialog(
     AlertDialog(
         text = {
             Text("Do you really want to delete the car \"${vehicle.name}\" ?\nThis action cannot be undone !")
-        },
-        onDismissRequest = onDismissRequest,
-        dismissButton = {
+        }, onDismissRequest = onDismissRequest, dismissButton = {
             TextButton(
                 onClick = onDismissRequest,
                 content = { Text("Cancel") },
                 modifier = Modifier.testTag(cancel)
             )
-        },
-        confirmButton = {
+        }, confirmButton = {
             TextButton(
                 onClick = onDelete,
                 content = { Text("Delete \"${vehicle.name}\"") },
                 modifier = Modifier.testTag(delete)
             )
-        },
-        modifier = modifier.testTag(DeleteVehicleButtonTags.Dialog.root)
+        }, modifier = modifier.testTag(DeleteVehicleButtonTags.Dialog.root)
     )
 }
